fix threading issues.
This issues was discovered when running unitetss. The program was
segfaulting due to threading issue. This change make sure we start and
close the sockets at the right place. Now we don't try to close anymore
the context, it is garbage collected at the end of the thread or the
program. Flapping, Trainer & Controller are also now sharing the context.