set stream backend instead of autodetecting it
Set stream_backend to geven or thread via the configuration instead of
detecting if gevent is installed and use it by default. It allows people
to still use thread event if gevent is here. By default the thread
backend is now used.

While I'm here refactor the stream module and split it in different
files to handle different imports.