make the configuration of redirection easier
Now the configuration of a stream can accept any class and any
options to allow more flexibilty.

Configuration in the ini should be passed in this form:

    [watcher:name]
    ...
    stream_stdout.class = FileStream
    stream_stdout.file = filename
    stream_stdout.refresh_time = 0.3

Same for `stream_stderr` .

if the `file` option is passed withut any `class`, the `FileStream`
class will be used by default.

By default the `refresh_time` is 0.3.

The watcher dict created from the ini is:

    {
        ..
        'stream_stdout': {
            'class': FileStream,
            'file': 'filename',
            'refresh_time': 0.3
        }
    }