Moved the flapping class in circus/plugins.
We have a CircusPlugin base class now that gives us the basic
wiring to write plugins: hook on the pub/sub of events,
and provide APIs to call commands against Circus.

The next step is to provide a way to activate and configure
plugins from the ini files and the API.

Then we can provide other plugins like a statsd emitter.