Add a FancyStdoutStream for color output
This takes some inspiration from honcho/foreman.

You can use this instead of the StdoutStream to provide colored output
with a similar prefix as circusd. If no color is selected a color will
be chosen at random.

Example:

  [watcher:foo]
  cmd = myserver
  stdout_stream.class = FancyStdoutStream
  stdout_stream.color = cyan

The available colors are red, green, yellow, blue, magenta, cyan,
and white.

In the future I can imagine being able to configure the prefix format.