Empty Includes
Currently, circus will fail to launch when one of the include directives
(include or include_dir) yields an empty set. This seems a little overly
protective and a simple warning emitted into the log should suffice.

Update _scan in read_config to write the previous exception message as a
log message at the warn level.