Add helper methods to CircusPlugin for parsing the received data
Before this change, each plug-in had to split data into "topic, msg", the u(topic).split('.'), then implicitly know that topic_parts[1] was the watcher name and topic_parts[2] was the action. Also, if the plug-in wanted to parse the msg portion, it needed to import zmq.utils.jsonapi. Not at all obvious to a plug-in writer.