no concurrent execution during arbiter restart
An explicit test is needed because the restart()
command is a kind of stop() command. The start after
the stop is made (externally) by circusd for plenty of reasons.

So the restart() command is over before the arbiter is really restarted.
So the classic synchronized decorator is released too early.

see #714