Add a --nostop option to the rm command
- This supports use cases for managing dynamically added processes
which run for a unknown amount of time. The processes can be managed
and respawned while they are running. However, when the processes
know they are done they should not be respawned anymore. The
processes (which are the only ones that know when they should be
completely stopped) can then tell circusd to remove them but not try
to kill them. The processes can then manage their own exit and not
have to worry about race conditions and syncronicity issues inherent
in trying to send a message to circusd which involves circusd trying
to kill the client sending the message.

- add a test for the nostop functionality