New option "close_child_stdin" for watchers
Circus, by default, always closes the "stdin" of the processes it forks.
While this might be good for a production setting, in development it is
not unusual to debug programs with

  import pdb; pdb.set_trace()

This will not work in processes spawned by Circus. Pdb raises a BdbQuit
exception on that instruction, since it can't receive user input.

The new option "close_child_stdin" is false, by default. This preserves
current Circus behaviour. Set it to true in a watcher configuration to
be able to attach an iteractive debugger session to its processes.