test_stdin_socket: Set use_sockets to True
In python 2.6, it appears that Popen() closes the stdin_socket
before preexec() is called, so when prexec() attempts to dup2()
the stdin_socket to fd 0, the old_fd  is already closed. It appears
that we can not allow the sockets to be closed in the Popen() call,
so setting use_sockets to True prevents that.