remove EAGAIN handler from Arbiter.reap_processes
try block in Arbiter.reap_processes contains a call
to os.waitpid and to Watcher.reap_process. Python's
os.waitpid does not throw OSError with errno set to
EAGAIN, functions called by Watcher.reap_process
also should not do that - so, there is no reason to
check for that errno value and handle it.