sort string representation of plugin config
Circus can compare string reprs of config to detect a
change (for example, it compares commands and stringified
config is a part of a command to launch plugin).
Converting to string is done by iterating over the dict,
which depends on the order of items in it, which can be
different for equal dicts. Sorting items of dict before
putting them to string will guarantee equal config
strings for equal configs.