test check_delay > 0 and call callback immediately
 - add test for positive check_delay

in production environment check_delay often set to a positive value, the
tornado API is changed so the PeriodicCallback no longer recieve event
loop as parameter.

 - use concurrent.future_add_done_callback which invoke callback immediately

When the future is already finished, we expect that the callback will be
executed immediately, but concurrent.add_done_callback did not make this
check and only add it to the event loop, while future_add_done_callback did this check.