Change ConfigParser to RawConfig parser to avoid interpolation errors
Closes #1097.

The first implementation of time_format feature was done in #493, there was also a test, which checked that the feature works as expected.

But the test was designed in such a way that it used FileStream class and not the configuration file (like some other tests). Therefore the error described in #1097 wouldn't appear.

I changed ConfigParser to RawConfigParser, because it seems that interpolation feature, which is specific to ConfigParser, is not widely used and is not described in the documentation. So there is no reason to use ConfigParser.

I added time_format to example2.ini and did not add more tests, because if it not breaks and all of the other tests are working, then it should be OK.