Fix failing resource watcher tests related to max_mem
Closes #1140.

I went through the test call-stack and found that the file which is used to test max memory is written before the the watcher is able to check the memory usage. Therefore the test fails.

The quickest solution is to add more chunks or increase chunk size so there will be enough time for the resource watcher to start and check system stats.

Playing with the chunk size is more of a workaround and not the actual fix, because the success rate of the test would highly depend on the machine, where tests are performed.