Add handling of garbage messages for controller socket
Fixes #1018 and closes #1019.

Disclaimer: all credit for this feature goes to @cgarciaarano. I am just expanding it a bit with a test.

The idea is that the "attacker" will be able to send malicious data over exposed controller socket and possibly overflow the machine and make it hang.

The test is straightforward: we are good if either dispatch or send_response methods were not called.