Dependencies: Add support for Python 3.10 (#1195)
Add Python 3.10 to the matrix of the CI workflow. All Python version
numbers are properly wrapped in strings, otherwise they are interpreted
as floats and `3.10` will become `3.1`.

The `setup.py` is updated to officially record support of Python 3.10
and the check for the minimum requirement is updated to actually check
for Python 3.7, it was mistakenly still checking for Python 3.5.

Finally, the tests for the `papa` optional dependency are disabled
because `papa` is not compatible with Python 3.10. It seems the project
has been abandoned as the latest officially supported version is Python
3.4 and the last release was in early 2016.