DevOps: Adopt PEP 621 and use `pyproject.toml` (#1197)
PEP 621 introduces the `pyproject.toml` file which is now the
recommended way for defining a project with its build and install
dependencies. In addition, it allows to provide configuration for
various tools, such as tox, instead of using individual files for each
tool.

The `MANIFEST.in` file is no longer required since the build backend
`flit` will automatically include all files, except those explicitly
marked with `[tool.flit.sdist] exclude`.

The description is added to the `__init__.py` of the package and it is
marked as `dynamic` in the `pyproject.toml` which will automatically
fetch it from the package docstring.