DevOps: Add continuous-deployment workflow
The `cd.yml` is a Github Actions workflow that triggers when a new tag
is pushed to the source repository. It will check that the version
specified in the `__version__` attribute matches that of the tag, and
if the case will run the full test suite. If all tests pass, the package
will be built using `flit` and pushed to PyPI. This will significantly
simplify and automate the release process.