Dependencies: Remove `papa` and related functionality
The `papa` library provides a process kernel that `circus` allowed ot
use through configuration. However, `papa` seems to be no longer
maintained as its last release was in January 2016. Officially it only
declares to be compatible up to Python 3.4. Apparently, so far it has
still been working until Python 3.10 where it actually breaks. Since it
is no longer maintained, we remove support for it from `circus`.