Add the `tests` path to the `PYTHONPATH` of the workers
The test scripts run by workers in the tests were importing modules from
the `circus/tests` directory. This was working since the `circus`
package would be automatically part of the `PYTHONPATH` when it was
installed.

Now that the `tests` directory is moved outside of the `circus` package
however, tests were failing because imports like `tests.generic` were no
longer importable. This is fixed by explicitly passing the `tests` path
in the `PYTHONPATH` environment variable that is passed to the workers
of the test arbiter.