Devops: Define the `__version__` attribute directly
The method to construct it from the `version_info` tuple causes the
`validate_release_tag.py` script used in the CD workflow to fail. The
abstract syntax tree parses stumbles on the `join` statement. The
following minimal example:

   ast.literal_eval(ast.parse("'.'.join([])").body[0])

raises the same exception:

    ValueError: malformed node or string on line 1

Since the `version_info` tuple is not actually used anywhere, the
problem is circumvented by simply defining the version string directly.